chore: move test_pagination in its own directory - DIA-46202
